Restructure packages and add SampledSignal to Signals
